trasform all images
